requirements.txt
flask
pydobotplus
pyserial
pyray